refactor(server): split runtime provision input
