+++ scripts/max_lines_allowlist.txt
@@ -36,7 +36,7 @@ server/proliferate/db/store/cloud_workspaces.py 1187 server structure migration 
 server/proliferate/server/billing/service.py 2459 server structure migration debt for oversized service file during billing orchestration consolidation
 server/proliferate/server/cloud/commands/domain/rules.py 607 server structure migration debt for oversized pure-domain file
 server/proliferate/server/cloud/mobility/service.py 1208 server structure migration debt for oversized service file
-server/proliferate/server/cloud/runtime/provision.py 975 server structure migration debt for oversized domain-adjacent file
+server/proliferate/server/cloud/runtime/provision.py 771 server structure migration debt for oversized domain-adjacent file
 server/proliferate/server/cloud/runtime_config/domain/resolver.py 611 server structure migration debt for oversized pure-domain file
 server/tests/integration/schema_migration_assertions.py 878 existing server oversized test helper
 server/tests/integration/test_auth_flow.py 2057 existing server oversized integration test
